feat: report controller

+++ mono/finance-app/src/main/java/com/example/finance_app/payload/CreateAccountRequest.java
@@ -7,7 +7,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -26,5 +25,5 @@ public class CreateAccountRequest {
 
     @NotNull
     @JsonProperty("balance")
-    private BigDecimal balance;
+    private double balance;
 }
